Add array_reverse*() and array_shift*() utilities.
Since realignment of the ring-buffer's underlying array was necessary, and
the most obvious way to do so was a circular shift, and I do not want to deal
with dynamic memory allocation, I implemented the reversal algorithm
described in Jon Bentley's "Programming Pearls 2nd edition", which is O(n) for
time and O(1) for space (i.e. the shift is done in-place).